updated 3sum code
