--- java/3Sum.java
@@ -1,37 +1,0 @@
-import java.util.*;
-
-public class ThreeSum {
-    public List<List<Integer>> threeSum(int[] nums) {
-        Arrays.sort(nums);
-        List<List<Integer>> result = new ArrayList<>();
-
-        for (int i = 0; i < nums.length - 2; i++) {
-            if (i > 0 && nums[i] == nums[i - 1]) continue; // Skip duplicates
-
-            int left = i + 1, right = nums.length - 1;
-            while (left < right) {
-                int sum = nums[i] + nums[left] + nums[right];
-
-                if (sum == 0) {
-                    result.add(Arrays.asList(nums[i], nums[left], nums[right]));
-                    while (left < right && nums[left] == nums[left + 1]) left++;
-                    while (left < right && nums[right] == nums[right - 1]) right--;
-                    left++;
-                    right--;
-                } else if (sum < 0) {
-                    left++;
-                } else {
-                    right--;
-                }
-            }
-        }
-
-        return result;
-    }
-
-    public static void main(String[] args) {
-        ThreeSum obj = new ThreeSum();
-        System.out.println(obj.threeSum(new int[]{-1,0,1,2,-1,-4}));
-        // [[-1,-1,2], [-1,0,1]]
-    }
-}
